Fix up namespaces of types move to ConnectedMode
Part of #4320
