+++ src/ConnectedMode.UnitTests/Binding/CredentialStoreTests.cs
@@ -24,7 +24,7 @@ using System.Linq.Expressions;
 using Microsoft.Alm.Authentication;
 using SonarLint.VisualStudio.TestInfrastructure;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests.LocalServices
+namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Binding
 {
     [TestClass]
     public class CredentialStoreTests
