+++ src/ConnectedMode.UnitTests/Persistence/SolutionBindingCredentialsLoaderTests.cs
@@ -20,12 +20,13 @@
 
 using System;
 using Microsoft.Alm.Authentication;
+using SonarLint.VisualStudio.ConnectedMode.Binding;
+using Moq;
 using SonarLint.VisualStudio.ConnectedMode.Persistence;
 using SonarLint.VisualStudio.Core.Binding;
-using SonarLint.VisualStudio.Integration.Persistence;
 using SonarQube.Client.Helpers;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests
+namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Persistence
 {
     [TestClass]
     public class SolutionBindingCredentialsLoaderTests
