+++ src/ConnectedMode/Persistence/ISolutionBindingCredentialsLoader.cs
@@ -21,7 +21,7 @@
 using System;
 using SonarLint.VisualStudio.Core.Binding;
 
-namespace SonarLint.VisualStudio.Integration.Persistence
+namespace SonarLint.VisualStudio.ConnectedMode.Persistence
 {
     interface ISolutionBindingCredentialsLoader
     {
