+++ src/ConnectedMode/Persistence/ISolutionBindingDataWriter.cs
@@ -20,9 +20,11 @@
 
 using System;
 using System.ComponentModel.Composition;
+using SonarLint.VisualStudio.ConnectedMode.Binding;
 using SonarLint.VisualStudio.Core.Binding;
+using SonarLint.VisualStudio.Integration;
 
-namespace SonarLint.VisualStudio.Integration.Persistence
+namespace SonarLint.VisualStudio.ConnectedMode.Persistence
 {
     internal interface ISolutionBindingDataWriter
     {
