+++ src/Integration/MefServices/ActiveSolutionBoundTracker.cs
@@ -26,7 +26,6 @@ using System.Threading.Tasks;
 using Microsoft.VisualStudio.Shell.Interop;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Core.Binding;
-using SonarLint.VisualStudio.Integration.NewConnectedMode;
 using SonarLint.VisualStudio.Integration.State;
 using SonarQube.Client;
